requirements.txt
rich
openai
subprocess
requests
urllib
bs4
chardet
typer